@@ -1,14 +1,22 @@
-import React from 'react';
+import { PropTypes } from 'react';
 import {GeoJson}  from 'react-leaflet';
 
 export default class AreaPolygon extends GeoJson {
 
 	componentWillReceiveProps(nextProps) {
+		if (nextProps.fillOpacity == 1) {
+			console.log(this.leafletElement);
+		}
 		if (nextProps.data !== this.props.data) {
 			this.leafletElement.clearLayers();
 		}
 		if (nextProps.className !== this.props.className) {
 			this.leafletElement.options.className = nextProps.className;
+		}
+
+		if (nextProps.fillOpacity !== this.props.fillOpacity) {
+			console.log(this.leafletElement);
+			this.leafletElement.options.fillOpacity = nextProps.fillOpacity;
 		}
 	}
 
@@ -16,9 +24,10 @@ export default class AreaPolygon extends GeoJson {
 		if (prevProps.data !== this.props.data) {
 			this.leafletElement.addData(this.props.data);
 		}
+		this.setStyleIfChanged(prevProps.style, this.props.style);
 	}
 }
 
 AreaPolygon.propTypes = {
-	data: React.PropTypes.object.isRequired
+	data: PropTypes.object.isRequired
 };
